Time: 363 ms (100.00%), Space: 110.7 MB (40.00%) - LeetHub

+++ 2412-minimum-money-required-before-transactions/2412-minimum-money-required-before-transactions.cpp
@@ -13,6 +13,6 @@ public:
             else maxa = max(maxa,t[i][0]);
         }
         
-        return loss + maxa;
+        return loss + (long long)maxa;
     }
 };
